+++ src/resqml2/ContinuousProperty.h
@@ -174,7 +174,7 @@ namespace RESQML2_NS
 		 * 										maximumValue and @p minimumValue then both minimum and
 		 *										maximum values will be computed from the values.
 		 */
-		DLL_IMPORT_OR_EXPORT void pushBackDoubleHdf5ArrayOfValues(double const * values, unsigned long long const * numValues, unsigned int numArrayDimensions,
+		DLL_IMPORT_OR_EXPORT void pushBackDoubleHdf5ArrayOfValues(double const * values, uint64_t const * numValues, unsigned int numArrayDimensions,
 			double * minimumValue, double * maximumValue, EML2_NS::AbstractHdfProxy* proxy = nullptr);
 		using AbstractValuesProperty::pushBackDoubleHdf5ArrayOfValues;
 
@@ -213,7 +213,7 @@ namespace RESQML2_NS
 		 * 									
 		 * @copydetails	pushBackDoubleHdf5ArrayOfValues
 		 */
-		DLL_IMPORT_OR_EXPORT void pushBackFloatHdf5ArrayOfValues(float const * values, unsigned long long const * numValues, unsigned int numArrayDimensions,
+		DLL_IMPORT_OR_EXPORT void pushBackFloatHdf5ArrayOfValues(float const * values, uint64_t const * numValues, unsigned int numArrayDimensions,
 			float * minimumValue, float * maximumValue, EML2_NS::AbstractHdfProxy* proxy = nullptr);
 		using AbstractValuesProperty::pushBackFloatHdf5ArrayOfValues;
 
@@ -298,7 +298,7 @@ namespace RESQML2_NS
 		 */
 		DLL_IMPORT_OR_EXPORT void pushBackHdf5ArrayOfValues(
 			COMMON_NS::AbstractObject::numericalDatatypeEnum datatype,
-			unsigned long long const * numValues,
+			uint64_t const * numValues,
 			unsigned int numArrayDimensions,
 			double * minimumValue, double * maximumValue,
 			EML2_NS::AbstractHdfProxy* proxy = nullptr
@@ -452,7 +452,7 @@ namespace RESQML2_NS
 			uint64_t offsetInSlowestDim,
 			bool computeMinMax,
 			EML2_NS::AbstractHdfProxy* proxy = nullptr,
-			unsigned int patchIndex = (std::numeric_limits<unsigned int>::max)()
+			uint64_t patchIndex = (std::numeric_limits<uint64_t>::max)()
 		);
 		using AbstractValuesProperty::setValuesOfFloatHdf5Array3dOfValues;
 		DLL_IMPORT_OR_EXPORT void setValuesOfDoubleHdf5Array3dOfValues(
@@ -465,7 +465,7 @@ namespace RESQML2_NS
 			uint64_t offsetInSlowestDim,
 			bool computeMinMax,
 			EML2_NS::AbstractHdfProxy* proxy = nullptr,
-			unsigned int patchIndex = (std::numeric_limits<unsigned int>::max)()
+			uint64_t patchIndex = (std::numeric_limits<uint64_t>::max)()
 		);
 		using AbstractValuesProperty::setValuesOfDoubleHdf5Array3dOfValues;
 
@@ -504,23 +504,23 @@ namespace RESQML2_NS
 		 */
 		DLL_IMPORT_OR_EXPORT void setValuesOfFloatHdf5ArrayOfValues(
 			float const * values,
-			unsigned long long const * numValues,
-			unsigned long long const * offsetValues,
+			uint64_t const * numValues,
+			uint64_t const * offsetValues,
 			unsigned int numArrayDimensions,
 			bool computeMinMax,
 			EML2_NS::AbstractHdfProxy* proxy = nullptr,
-			unsigned int patchIndex = (std::numeric_limits<unsigned int>::max)()
+			uint64_t patchIndex = (std::numeric_limits<uint64_t>::max)()
 		);
 		using AbstractValuesProperty::setValuesOfFloatHdf5ArrayOfValues;
 		
 		DLL_IMPORT_OR_EXPORT void setValuesOfDoubleHdf5ArrayOfValues(
 			double const * values,
-			unsigned long long const * numValues,
-			unsigned long long const * offsetValues,
+			uint64_t const * numValues,
+			uint64_t const * offsetValues,
 			unsigned int numArrayDimensions,
 			bool computeMinMax,
 			EML2_NS::AbstractHdfProxy* proxy = nullptr,
-			unsigned int patchIndex = (std::numeric_limits<unsigned int>::max)()
+			uint64_t patchIndex = (std::numeric_limits<uint64_t>::max)()
 		);
 		using AbstractValuesProperty::setValuesOfDoubleHdf5ArrayOfValues;
 
@@ -579,7 +579,7 @@ namespace RESQML2_NS
 		template <class T>
 		void setPropertyMinMax(
 			T const * values, 
-			unsigned long long const * numValuesInEachDimension,
+			uint64_t const * numValuesInEachDimension,
 			unsigned int numArrayDimensions,
 			T * minimumValue = nullptr, T * maximumValue = nullptr)
 		{
