+++ src/resqml2_0_1/PointsProperty.cpp
@@ -87,7 +87,7 @@ EML2_NS::AbstractHdfProxy * PointsProperty::getDatasetOfPatch(uint64_t patchInde
 
 	gsoap_resqml2_0_1::resqml20__PatchOfPoints* patch = static_cast<_resqml20__PointsProperty*>(gsoapProxy2_0_1)->PatchOfPoints[patchIndex];
 
-	nullValue = (numeric_limits<long>::min)();
+	nullValue = (numeric_limits<int64_t>::min)();
 	int valuesType = patch->Points->soap_type();
 	if (valuesType == SOAP_TYPE_gsoap_resqml2_0_1_resqml20__Point3dHdf5Array) {
 		dsPath = static_cast<gsoap_resqml2_0_1::resqml20__Point3dHdf5Array*>(patch->Points)->Coordinates->PathInHdfFile;
@@ -98,7 +98,7 @@ EML2_NS::AbstractHdfProxy * PointsProperty::getDatasetOfPatch(uint64_t patchInde
 	}
 }
 
-COMMON_NS::DataObjectReference PointsProperty::getHdfProxyDor(unsigned int patchIndex) const
+COMMON_NS::DataObjectReference PointsProperty::getHdfProxyDor(uint64_t patchIndex) const
 {
 	if (patchIndex >= getPatchCount()) {
 		throw out_of_range("The values property patch is out of range");
