+++ src/resqml2_0_1/PointsProperty.h
@@ -151,7 +151,7 @@ namespace RESQML2_0_1_NS
 		 *
 		 * @param 		  	patchIndex	The corresponding patch index of the dataset to get.
 		 * @param [out]		nullValue 	If possible, this function will set this parameter to the RESQML
-		 * 								null value of the dataset. If not, it will return long.min.
+		 * 								null value of the dataset. If not, it will return int64_t.min.
 		 * @param [out]		nullValue 	this function will set this parameter to the RESQML
 		 * 								dataset path in the HDF file.
 		 *
@@ -168,7 +168,7 @@ namespace RESQML2_0_1_NS
 		 * 			Otherwise returns the data object reference of the associated representation. Null
 		 * 			should not occured since each property must be associated to a representation.
 		 */
-		COMMON_NS::DataObjectReference getHdfProxyDor(unsigned int patchIndex) const final;
+		COMMON_NS::DataObjectReference getHdfProxyDor(uint64_t patchIndex) const final;
 
 	private:
 		EML2_NS::AbstractHdfProxy* getValuesHdfProxyAndDatasetPathOfPatch(unsigned int patchIndex, std::string & datasetPath) const final;
